pyproject.toml
[project]
name = "dj-events-face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.8",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "ruff>=0.14.5",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
